+++ src/eth_client/libethereum/ExtVM.h
@@ -82,7 +82,7 @@ public:
     }
 
     /// Selfdestruct the associated contract to the given address.
-    void selfdestruct(Address _a) final;
+    virtual bool selfdestruct(Address _a) final;
 
     /// Return the EVM gas-price schedule for this execution context.
     EVMSchedule const& evmSchedule() const final { return m_evmSchedule; }
