+++ src/eth_client/libevm/ExtVMFace.cpp
@@ -41,7 +41,7 @@ evmc_storage_status EvmCHost::set_storage(
     u256 const currentValue = m_extVM.store(index);
 
     if (newValue == currentValue)
-        return EVMC_STORAGE_UNCHANGED;
+        return EVMC_STORAGE_ASSIGNED;
 
     EVMSchedule const& schedule = m_extVM.evmSchedule();
     auto status = EVMC_STORAGE_MODIFIED;
@@ -58,7 +58,7 @@ evmc_storage_status EvmCHost::set_storage(
     }
     else
     {
-        status = EVMC_STORAGE_MODIFIED_AGAIN;
+        status = EVMC_STORAGE_ASSIGNED;
         if (originalValue != 0)
         {
             if (currentValue == 0)
@@ -115,11 +115,11 @@ size_t EvmCHost::copy_code(evmc::address const& _addr, size_t _codeOffset, byte*
     return numToCopy;
 }
 
-void EvmCHost::selfdestruct(evmc::address const& _addr, evmc::address const& _beneficiary) noexcept
+bool EvmCHost::selfdestruct(evmc::address const& _addr, evmc::address const& _beneficiary) noexcept
 {
     assert(fromEvmC(_addr) == m_extVM.myAddress);
     record_account_access(_addr);
-    m_extVM.selfdestruct(fromEvmC(_beneficiary));
+    return m_extVM.selfdestruct(fromEvmC(_beneficiary));
 }
 
 
@@ -173,7 +173,7 @@ evmc_tx_context EvmCHost::get_tx_context() const noexcept
     result.block_number = envInfo.number();
     result.block_timestamp = envInfo.timestamp();
     result.block_gas_limit = static_cast<int64_t>(envInfo.gasLimit());
-    result.block_difficulty = toEvmC(envInfo.difficulty());
+    result.block_prev_randao = toEvmC(envInfo.difficulty());
     result.chain_id = toEvmC(envInfo.chainID());
     return result;
 }
@@ -183,7 +183,7 @@ evmc::bytes32 EvmCHost::get_block_hash(int64_t _number) const noexcept
     return toEvmC(m_extVM.blockHash(_number));
 }
 
-evmc::result EvmCHost::create(evmc_message const& _msg) noexcept
+evmc::Result EvmCHost::create(evmc_message const& _msg) noexcept
 {
     u256 gas = _msg.gas;
     u256 value = fromEvmC(_msg.value);
@@ -224,15 +224,15 @@ evmc::result EvmCHost::create(evmc_message const& _msg) noexcept
             output.~bytes();
         };
     }
-    return evmc::result{evmcResult};
+    return evmc::Result{evmcResult};
 }
 
-evmc::result EvmCHost::call(evmc_message const& _msg) noexcept
+evmc::Result EvmCHost::call(evmc_message const& _msg) noexcept
 {
     assert(_msg.gas >= 0 && "Invalid gas value");
     assert(_msg.depth == static_cast<int>(m_extVM.depth) + 1);
 
-    record_account_access(_msg.destination);
+    record_account_access(_msg.recipient);
 
     // Handle CREATE separately.
     if (_msg.kind == EVMC_CREATE || _msg.kind == EVMC_CREATE2)
@@ -243,7 +243,7 @@ evmc::result EvmCHost::call(evmc_message const& _msg) noexcept
     params.apparentValue = fromEvmC(_msg.value);
     params.valueTransfer = _msg.kind == EVMC_DELEGATECALL ? 0 : params.apparentValue;
     params.senderAddress = fromEvmC(_msg.sender);
-    params.codeAddress = fromEvmC(_msg.destination);
+    params.codeAddress = fromEvmC(_msg.recipient);
     params.receiveAddress = _msg.kind == EVMC_CALL ? params.codeAddress : m_extVM.myAddress;
     params.data = {_msg.input_data, _msg.input_size};
     params.staticCall = (_msg.flags & EVMC_STATIC) != 0;
@@ -274,7 +274,7 @@ evmc::result EvmCHost::call(evmc_message const& _msg) noexcept
         // This is normal pattern when placement new operator is used.
         output.~bytes();
     };
-    return evmc::result{evmcResult};
+    return evmc::Result{evmcResult};
 }
 
 ExtVMFace::ExtVMFace(EnvInfo const& _envInfo, Address _myAddress, Address _caller, Address _origin,
