+++ src/eth_client/utils/intx/intx.hpp
@@ -1,19 +1,1038 @@
 // intx: extended precision integer library.
-// Copyright 2019-2020 Pawel Bylica.
+// Copyright 2019 Pawel Bylica.
 // Licensed under the Apache License, Version 2.0.
 
 #pragma once
 
-#include <intx/int128.hpp>
 #include <algorithm>
-#include <array>
+#include <cassert>
+#include <climits>
 #include <cstdint>
 #include <cstring>
 #include <limits>
+#include <stdexcept>
+#include <string>
+#include <tuple>
 #include <type_traits>
+
+#ifdef _MSC_VER
+    #pragma warning(push)
+    #pragma warning(disable : 5030)  // Allow unknown attributes.
+#endif
+
+
+#ifndef __has_builtin
+    #define __has_builtin(NAME) 0
+#endif
+
+#ifdef _MSC_VER
+    #include <intrin.h>
+#endif
+
+#if !defined(__has_builtin)
+    #define __has_builtin(NAME) 0
+#endif
+
+#if !defined(__has_feature)
+    #define __has_feature(NAME) 0
+#endif
+
+#if !defined(NDEBUG)
+    #define INTX_UNREACHABLE() assert(false)
+#elif __has_builtin(__builtin_unreachable)
+    #define INTX_UNREACHABLE() __builtin_unreachable()
+#elif defined(_MSC_VER)
+    #define INTX_UNREACHABLE() __assume(0)
+#else
+    #define INTX_UNREACHABLE() (void)0
+#endif
+
+
+#if __has_builtin(__builtin_expect)
+    #define INTX_UNLIKELY(EXPR) __builtin_expect(bool{EXPR}, false)
+#else
+    #define INTX_UNLIKELY(EXPR) (bool{EXPR})
+#endif
+
+#if !defined(NDEBUG)
+    #define INTX_REQUIRE assert
+#else
+    #define INTX_REQUIRE(X) (X) ? (void)0 : INTX_UNREACHABLE()
+#endif
+
+
+// Detect compiler support for 128-bit integer __int128
+#if defined(__SIZEOF_INT128__)
+    #define INTX_HAS_BUILTIN_INT128 1
+#else
+    #define INTX_HAS_BUILTIN_INT128 0
+#endif
 
 namespace intx
 {
+/// Mark a possible code path as unreachable (invokes undefined behavior).
+/// TODO(C++23): Use std::unreachable().
+[[noreturn]] inline void unreachable() noexcept
+{
+#if __has_builtin(__builtin_unreachable)
+    __builtin_unreachable();
+#elif defined(_MSC_VER)
+    __assume(false);
+#endif
+}
+
+#if INTX_HAS_BUILTIN_INT128
+    #pragma GCC diagnostic push
+    #pragma GCC diagnostic ignored "-Wpedantic"  // Usage of __int128 triggers a pedantic warning.
+
+/// Alias for the compiler supported unsigned __int128 type.
+using builtin_uint128 = unsigned __int128;
+
+    #pragma GCC diagnostic pop
+#endif
+
+constexpr bool byte_order_is_little_endian =
+#if defined(__BYTE_ORDER__)
+    (__BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__);
+#elif defined(_WIN32)
+    true;  // On Windows assume little endian.
+#else
+    #error "Unknown endianness"
+#endif
+
+template <unsigned N>
+struct uint;
+
+/// The 128-bit unsigned integer.
+///
+/// This type is defined as a specialization of uint<> to easier integration with full intx package,
+/// however, uint128 may be used independently.
+template <>
+struct uint<128>
+{
+    using word_type = uint64_t;
+    static constexpr auto word_num_bits = sizeof(word_type) * 8;
+    static constexpr unsigned num_bits = 128;
+    static constexpr auto num_words = num_bits / word_num_bits;
+
+private:
+    uint64_t words_[2]{};
+
+public:
+    constexpr uint() noexcept = default;
+
+    constexpr uint(uint64_t low, uint64_t high) noexcept : words_{low, high} {}
+
+    template <typename T,
+        typename = typename std::enable_if_t<std::is_convertible<T, uint64_t>::value>>
+    constexpr uint(T x) noexcept : words_{static_cast<uint64_t>(x), 0}  // NOLINT
+    {}
+
+#if INTX_HAS_BUILTIN_INT128
+    constexpr uint(builtin_uint128 x) noexcept  // NOLINT
+      : words_{uint64_t(x), uint64_t(x >> 64)}
+    {}
+
+    constexpr explicit operator builtin_uint128() const noexcept
+    {
+        return (builtin_uint128{words_[1]} << 64) | words_[0];
+    }
+#endif
+
+    constexpr uint64_t& operator[](size_t i) noexcept { return words_[i]; }
+    constexpr const uint64_t& operator[](size_t i) const noexcept { return words_[i]; }
+
+    constexpr explicit operator bool() const noexcept { return (words_[0] | words_[1]) != 0; }
+
+    /// Explicit converting operator for all builtin integral types.
+    template <typename Int, typename = typename std::enable_if<std::is_integral<Int>::value>::type>
+    constexpr explicit operator Int() const noexcept
+    {
+        return static_cast<Int>(words_[0]);
+    }
+};
+
+using uint128 = uint<128>;
+
+
+inline constexpr bool is_constant_evaluated() noexcept
+{
+#if __has_builtin(__builtin_is_constant_evaluated) || (defined(_MSC_VER) && _MSC_VER >= 1925)
+    return __builtin_is_constant_evaluated();
+#else
+    return true;
+#endif
+}
+
+
+/// Contains result of add/sub/etc with a carry flag.
+template <typename T>
+struct result_with_carry
+{
+    T value;
+    bool carry;
+
+    /// Conversion to tuple of references, to allow usage with std::tie().
+    constexpr operator std::tuple<T&, bool&>() noexcept { return {value, carry}; }
+};
+
+
+/// Linear arithmetic operators.
+/// @{
+
+/// Addition with carry.
+inline constexpr result_with_carry<uint64_t> addc(
+    uint64_t x, uint64_t y, bool carry = false) noexcept
+{
+#if __has_builtin(__builtin_addcll)
+    if (!is_constant_evaluated())
+    {
+        unsigned long long carryout = 0;  // NOLINT(google-runtime-int)
+        const auto s = __builtin_addcll(x, y, carry, &carryout);
+        static_assert(sizeof(s) == sizeof(uint64_t));
+        return {s, static_cast<bool>(carryout)};
+    }
+#elif __has_builtin(__builtin_ia32_addcarryx_u64)
+    if (!is_constant_evaluated())
+    {
+        unsigned long long s = 0;  // NOLINT(google-runtime-int)
+        static_assert(sizeof(s) == sizeof(uint64_t));
+        const auto carryout = __builtin_ia32_addcarryx_u64(carry, x, y, &s);
+        return {s, static_cast<bool>(carryout)};
+    }
+#endif
+
+    const auto s = x + y;
+    const auto carry1 = s < x;
+    const auto t = s + carry;
+    const auto carry2 = t < s;
+    return {t, carry1 || carry2};
+}
+
+/// Subtraction with carry (borrow).
+inline constexpr result_with_carry<uint64_t> subc(
+    uint64_t x, uint64_t y, bool carry = false) noexcept
+{
+#if __has_builtin(__builtin_subcll)
+    if (!is_constant_evaluated())
+    {
+        unsigned long long carryout = 0;  // NOLINT(google-runtime-int)
+        const auto d = __builtin_subcll(x, y, carry, &carryout);
+        static_assert(sizeof(d) == sizeof(uint64_t));
+        return {d, static_cast<bool>(carryout)};
+    }
+#elif __has_builtin(__builtin_ia32_sbb_u64)
+    if (!is_constant_evaluated())
+    {
+        unsigned long long d = 0;  // NOLINT(google-runtime-int)
+        static_assert(sizeof(d) == sizeof(uint64_t));
+        const auto carryout = __builtin_ia32_sbb_u64(carry, x, y, &d);
+        return {d, static_cast<bool>(carryout)};
+    }
+#endif
+
+    const auto d = x - y;
+    const auto carry1 = x < y;
+    const auto e = d - carry;
+    const auto carry2 = d < uint64_t{carry};
+    return {e, carry1 || carry2};
+}
+
+/// Addition with carry.
+template <unsigned N>
+inline constexpr result_with_carry<uint<N>> addc(
+    const uint<N>& x, const uint<N>& y, bool carry = false) noexcept
+{
+    uint<N> s;
+    bool k = carry;
+    for (size_t i = 0; i < uint<N>::num_words; ++i)
+    {
+        auto t = addc(x[i], y[i], k);
+        s[i] = t.value;
+        k = t.carry;
+    }
+    return {s, k};
+}
+
+inline constexpr uint128 operator+(uint128 x, uint128 y) noexcept
+{
+    return addc(x, y).value;
+}
+
+inline constexpr uint128 operator+(uint128 x) noexcept
+{
+    return x;
+}
+
+/// Performs subtraction of two unsigned numbers and returns the difference
+/// and the carry bit (aka borrow, overflow).
+template <unsigned N>
+inline constexpr result_with_carry<uint<N>> subc(
+    const uint<N>& x, const uint<N>& y, bool carry = false) noexcept
+{
+    uint<N> z;
+    bool k = carry;
+    for (size_t i = 0; i < uint<N>::num_words; ++i)
+    {
+        auto t = subc(x[i], y[i], k);
+        z[i] = t.value;
+        k = t.carry;
+    }
+    return {z, k};
+}
+
+inline constexpr uint128 operator-(uint128 x, uint128 y) noexcept
+{
+    return subc(x, y).value;
+}
+
+inline constexpr uint128 operator-(uint128 x) noexcept
+{
+    // Implementing as subtraction is better than ~x + 1.
+    // Clang9: Perfect.
+    // GCC8: Does something weird.
+    return 0 - x;
+}
+
+inline uint128& operator++(uint128& x) noexcept
+{
+    return x = x + 1;
+}
+
+inline uint128& operator--(uint128& x) noexcept
+{
+    return x = x - 1;
+}
+
+inline const uint128 operator++(uint128& x, int) noexcept  // NOLINT(readability-const-return-type)
+{
+    const auto ret = x;
+    ++x;
+    return ret;
+}
+
+inline const uint128 operator--(uint128& x, int) noexcept  // NOLINT(readability-const-return-type)
+{
+    const auto ret = x;
+    --x;
+    return ret;
+}
+
+/// Optimized addition.
+///
+/// This keeps the multiprecision addition until CodeGen so the pattern is not
+/// broken during other optimizations.
+inline constexpr uint128 fast_add(uint128 x, uint128 y) noexcept
+{
+#if INTX_HAS_BUILTIN_INT128
+    return builtin_uint128{x} + builtin_uint128{y};
+#else
+    return x + y;  // Fallback to generic addition.
+#endif
+}
+
+/// @}
+
+
+/// Comparison operators.
+///
+/// In all implementations bitwise operators are used instead of logical ones
+/// to avoid branching.
+///
+/// @{
+
+inline constexpr bool operator==(uint128 x, uint128 y) noexcept
+{
+    return ((x[0] ^ y[0]) | (x[1] ^ y[1])) == 0;
+}
+
+inline constexpr bool operator!=(uint128 x, uint128 y) noexcept
+{
+    return !(x == y);
+}
+
+inline constexpr bool operator<(uint128 x, uint128 y) noexcept
+{
+    // OPT: This should be implemented by checking the borrow of x - y,
+    //      but compilers (GCC8, Clang7)
+    //      have problem with properly optimizing subtraction.
+#if INTX_HAS_BUILTIN_INT128
+    return builtin_uint128{x} < builtin_uint128{y};
+#else
+    return (unsigned{x[1] < y[1]} | (unsigned{x[1] == y[1]} & unsigned{x[0] < y[0]})) != 0;
+#endif
+}
+
+inline constexpr bool operator<=(uint128 x, uint128 y) noexcept
+{
+    return !(y < x);
+}
+
+inline constexpr bool operator>(uint128 x, uint128 y) noexcept
+{
+    return y < x;
+}
+
+inline constexpr bool operator>=(uint128 x, uint128 y) noexcept
+{
+    return !(x < y);
+}
+
+/// @}
+
+
+/// Bitwise operators.
+/// @{
+
+inline constexpr uint128 operator~(uint128 x) noexcept
+{
+    return {~x[0], ~x[1]};
+}
+
+inline constexpr uint128 operator|(uint128 x, uint128 y) noexcept
+{
+    // Clang: perfect.
+    // GCC 8-12: stupidly uses a vector instruction in all bitwise operators.
+    return {x[0] | y[0], x[1] | y[1]};
+}
+
+inline constexpr uint128 operator&(uint128 x, uint128 y) noexcept
+{
+    return {x[0] & y[0], x[1] & y[1]};
+}
+
+inline constexpr uint128 operator^(uint128 x, uint128 y) noexcept
+{
+    return {x[0] ^ y[0], x[1] ^ y[1]};
+}
+
+inline constexpr uint128 operator<<(uint128 x, uint64_t shift) noexcept
+{
+    return (shift < 64) ?
+               // Find the part moved from lo to hi.
+               // For shift == 0 right shift by (64 - shift) is invalid so
+               // split it into 2 shifts by 1 and (63 - shift).
+               uint128{x[0] << shift, (x[1] << shift) | ((x[0] >> 1) >> (63 - shift))} :
+
+               // Guarantee "defined" behavior for shifts larger than 128.
+               (shift < 128) ? uint128{0, x[0] << (shift - 64)} : 0;
+}
+
+inline constexpr uint128 operator<<(uint128 x, uint128 shift) noexcept
+{
+    if (INTX_UNLIKELY(shift[1] != 0))
+        return 0;
+
+    return x << shift[0];
+}
+
+inline constexpr uint128 operator>>(uint128 x, uint64_t shift) noexcept
+{
+    return (shift < 64) ?
+               // Find the part moved from lo to hi.
+               // For shift == 0 left shift by (64 - shift) is invalid so
+               // split it into 2 shifts by 1 and (63 - shift).
+               uint128{(x[0] >> shift) | ((x[1] << 1) << (63 - shift)), x[1] >> shift} :
+
+               // Guarantee "defined" behavior for shifts larger than 128.
+               (shift < 128) ? uint128{x[1] >> (shift - 64)} : 0;
+}
+
+inline constexpr uint128 operator>>(uint128 x, uint128 shift) noexcept
+{
+    if (INTX_UNLIKELY(shift[1] != 0))
+        return 0;
+
+    return x >> static_cast<uint64_t>(shift);
+}
+
+/// @}
+
+
+/// Multiplication
+/// @{
+
+/// Full unsigned multiplication 64 x 64 -> 128.
+inline constexpr uint128 umul(uint64_t x, uint64_t y) noexcept
+{
+#if INTX_HAS_BUILTIN_INT128
+    return builtin_uint128{x} * builtin_uint128{y};
+#elif defined(_MSC_VER) && _MSC_VER >= 1925 && defined(_M_X64)
+    if (!is_constant_evaluated())
+    {
+        unsigned __int64 hi = 0;
+        const auto lo = _umul128(x, y, &hi);
+        return {lo, hi};
+    }
+    // For constexpr fallback to portable variant.
+#endif
+
+    // Portable full unsigned multiplication 64 x 64 -> 128.
+    uint64_t xl = x & 0xffffffff;
+    uint64_t xh = x >> 32;
+    uint64_t yl = y & 0xffffffff;
+    uint64_t yh = y >> 32;
+
+    uint64_t t0 = xl * yl;
+    uint64_t t1 = xh * yl;
+    uint64_t t2 = xl * yh;
+    uint64_t t3 = xh * yh;
+
+    uint64_t u1 = t1 + (t0 >> 32);
+    uint64_t u2 = t2 + (u1 & 0xffffffff);
+
+    uint64_t lo = (u2 << 32) | (t0 & 0xffffffff);
+    uint64_t hi = t3 + (u2 >> 32) + (u1 >> 32);
+    return {lo, hi};
+}
+
+inline constexpr uint128 operator*(uint128 x, uint128 y) noexcept
+{
+    auto p = umul(x[0], y[0]);
+    p[1] += (x[0] * y[1]) + (x[1] * y[0]);
+    return {p[0], p[1]};
+}
+
+/// @}
+
+
+/// Assignment operators.
+/// @{
+
+inline constexpr uint128& operator+=(uint128& x, uint128 y) noexcept
+{
+    return x = x + y;
+}
+
+inline constexpr uint128& operator-=(uint128& x, uint128 y) noexcept
+{
+    return x = x - y;
+}
+
+inline uint128& operator*=(uint128& x, uint128 y) noexcept
+{
+    return x = x * y;
+}
+
+inline constexpr uint128& operator|=(uint128& x, uint128 y) noexcept
+{
+    return x = x | y;
+}
+
+inline constexpr uint128& operator&=(uint128& x, uint128 y) noexcept
+{
+    return x = x & y;
+}
+
+inline constexpr uint128& operator^=(uint128& x, uint128 y) noexcept
+{
+    return x = x ^ y;
+}
+
+inline constexpr uint128& operator<<=(uint128& x, uint64_t shift) noexcept
+{
+    return x = x << shift;
+}
+
+inline constexpr uint128& operator>>=(uint128& x, uint64_t shift) noexcept
+{
+    return x = x >> shift;
+}
+
+/// @}
+
+
+inline constexpr unsigned clz_generic(uint32_t x) noexcept
+{
+    unsigned n = 32;
+    for (int i = 4; i >= 0; --i)
+    {
+        const auto s = unsigned{1} << i;
+        const auto hi = x >> s;
+        if (hi != 0)
+        {
+            n -= s;
+            x = hi;
+        }
+    }
+    return n - x;
+}
+
+inline constexpr unsigned clz_generic(uint64_t x) noexcept
+{
+    unsigned n = 64;
+    for (int i = 5; i >= 0; --i)
+    {
+        const auto s = unsigned{1} << i;
+        const auto hi = x >> s;
+        if (hi != 0)
+        {
+            n -= s;
+            x = hi;
+        }
+    }
+    return n - static_cast<unsigned>(x);
+}
+
+inline constexpr unsigned clz(uint32_t x) noexcept
+{
+#ifdef _MSC_VER
+    return clz_generic(x);
+#else
+    return x != 0 ? unsigned(__builtin_clz(x)) : 32;
+#endif
+}
+
+inline constexpr unsigned clz(uint64_t x) noexcept
+{
+#ifdef _MSC_VER
+    return clz_generic(x);
+#else
+    return x != 0 ? unsigned(__builtin_clzll(x)) : 64;
+#endif
+}
+
+inline constexpr unsigned clz(uint128 x) noexcept
+{
+    // In this order `h == 0` we get less instructions than in case of `h != 0`.
+    return x[1] == 0 ? clz(x[0]) + 64 : clz(x[1]);
+}
+
+template <typename T>
+T bswap(T x) noexcept = delete;  // Disable type auto promotion
+
+inline constexpr uint8_t bswap(uint8_t x) noexcept
+{
+    return x;
+}
+
+inline constexpr uint16_t bswap(uint16_t x) noexcept
+{
+#if __has_builtin(__builtin_bswap16)
+    return __builtin_bswap16(x);
+#else
+    #ifdef _MSC_VER
+    if (!is_constant_evaluated())
+        return _byteswap_ushort(x);
+    #endif
+    return static_cast<uint16_t>((x << 8) | (x >> 8));
+#endif
+}
+
+inline constexpr uint32_t bswap(uint32_t x) noexcept
+{
+#if __has_builtin(__builtin_bswap32)
+    return __builtin_bswap32(x);
+#else
+    #ifdef _MSC_VER
+    if (!is_constant_evaluated())
+        return _byteswap_ulong(x);
+    #endif
+    const auto a = ((x << 8) & 0xFF00FF00) | ((x >> 8) & 0x00FF00FF);
+    return (a << 16) | (a >> 16);
+#endif
+}
+
+inline constexpr uint64_t bswap(uint64_t x) noexcept
+{
+#if __has_builtin(__builtin_bswap64)
+    return __builtin_bswap64(x);
+#else
+    #ifdef _MSC_VER
+    if (!is_constant_evaluated())
+        return _byteswap_uint64(x);
+    #endif
+    const auto a = ((x << 8) & 0xFF00FF00FF00FF00) | ((x >> 8) & 0x00FF00FF00FF00FF);
+    const auto b = ((a << 16) & 0xFFFF0000FFFF0000) | ((a >> 16) & 0x0000FFFF0000FFFF);
+    return (b << 32) | (b >> 32);
+#endif
+}
+
+inline constexpr uint128 bswap(uint128 x) noexcept
+{
+    return {bswap(x[1]), bswap(x[0])};
+}
+
+
+/// Division.
+/// @{
+
+template <typename QuotT, typename RemT = QuotT>
+struct div_result
+{
+    QuotT quot;
+    RemT rem;
+
+    /// Conversion to tuple of references, to allow usage with std::tie().
+    constexpr operator std::tuple<QuotT&, RemT&>() noexcept { return {quot, rem}; }
+};
+
+namespace internal
+{
+inline constexpr uint16_t reciprocal_table_item(uint8_t d9) noexcept
+{
+    return uint16_t(0x7fd00 / (0x100 | d9));
+}
+
+#define REPEAT4(x)                                                  \
+    reciprocal_table_item((x) + 0), reciprocal_table_item((x) + 1), \
+        reciprocal_table_item((x) + 2), reciprocal_table_item((x) + 3)
+
+#define REPEAT32(x)                                                                         \
+    REPEAT4((x) + 4 * 0), REPEAT4((x) + 4 * 1), REPEAT4((x) + 4 * 2), REPEAT4((x) + 4 * 3), \
+        REPEAT4((x) + 4 * 4), REPEAT4((x) + 4 * 5), REPEAT4((x) + 4 * 6), REPEAT4((x) + 4 * 7)
+
+#define REPEAT256()                                                                           \
+    REPEAT32(32 * 0), REPEAT32(32 * 1), REPEAT32(32 * 2), REPEAT32(32 * 3), REPEAT32(32 * 4), \
+        REPEAT32(32 * 5), REPEAT32(32 * 6), REPEAT32(32 * 7)
+
+/// Reciprocal lookup table.
+constexpr uint16_t reciprocal_table[] = {REPEAT256()};
+
+#undef REPEAT4
+#undef REPEAT32
+#undef REPEAT256
+}  // namespace internal
+
+/// Computes the reciprocal (2^128 - 1) / d - 2^64 for normalized d.
+///
+/// Based on Algorithm 2 from "Improved division by invariant integers".
+inline uint64_t reciprocal_2by1(uint64_t d) noexcept
+{
+    INTX_REQUIRE(d & 0x8000000000000000);  // Must be normalized.
+
+    const uint64_t d9 = d >> 55;
+    const uint32_t v0 = internal::reciprocal_table[d9 - 256];
+
+    const uint64_t d40 = (d >> 24) + 1;
+    const uint64_t v1 = (v0 << 11) - uint32_t(uint32_t{v0 * v0} * d40 >> 40) - 1;
+
+    const uint64_t v2 = (v1 << 13) + (v1 * (0x1000000000000000 - v1 * d40) >> 47);
+
+    const uint64_t d0 = d & 1;
+    const uint64_t d63 = (d >> 1) + d0;  // ceil(d/2)
+    const uint64_t e = ((v2 >> 1) & (0 - d0)) - v2 * d63;
+    const uint64_t v3 = (umul(v2, e)[1] >> 1) + (v2 << 31);
+
+    const uint64_t v4 = v3 - (umul(v3, d) + d)[1] - d;
+    return v4;
+}
+
+inline uint64_t reciprocal_3by2(uint128 d) noexcept
+{
+    auto v = reciprocal_2by1(d[1]);
+    auto p = d[1] * v;
+    p += d[0];
+    if (p < d[0])
+    {
+        --v;
+        if (p >= d[1])
+        {
+            --v;
+            p -= d[1];
+        }
+        p -= d[1];
+    }
+
+    const auto t = umul(v, d[0]);
+
+    p += t[1];
+    if (p < t[1])
+    {
+        --v;
+        if (p >= d[1])
+        {
+            if (p > d[1] || t[0] >= d[0])
+                --v;
+        }
+    }
+    return v;
+}
+
+inline div_result<uint64_t> udivrem_2by1(uint128 u, uint64_t d, uint64_t v) noexcept
+{
+    auto q = umul(v, u[1]);
+    q = fast_add(q, u);
+
+    ++q[1];
+
+    auto r = u[0] - q[1] * d;
+
+    if (r > q[0])
+    {
+        --q[1];
+        r += d;
+    }
+
+    if (r >= d)
+    {
+        ++q[1];
+        r -= d;
+    }
+
+    return {q[1], r};
+}
+
+inline div_result<uint64_t, uint128> udivrem_3by2(
+    uint64_t u2, uint64_t u1, uint64_t u0, uint128 d, uint64_t v) noexcept
+{
+    auto q = umul(v, u2);
+    q = fast_add(q, {u1, u2});
+
+    auto r1 = u1 - q[1] * d[1];
+
+    auto t = umul(d[0], q[1]);
+
+    auto r = uint128{u0, r1} - t - d;
+    r1 = r[1];
+
+    ++q[1];
+
+    if (r1 >= q[0])
+    {
+        --q[1];
+        r += d;
+    }
+
+    if (r >= d)
+    {
+        ++q[1];
+        r -= d;
+    }
+
+    return {q[1], r};
+}
+
+inline div_result<uint128> udivrem(uint128 x, uint128 y) noexcept
+{
+    if (y[1] == 0)
+    {
+        INTX_REQUIRE(y[0] != 0);  // Division by 0.
+
+        const auto lsh = clz(y[0]);
+        const auto rsh = (64 - lsh) % 64;
+        const auto rsh_mask = uint64_t{lsh == 0} - 1;
+
+        const auto yn = y[0] << lsh;
+        const auto xn_lo = x[0] << lsh;
+        const auto xn_hi = (x[1] << lsh) | ((x[0] >> rsh) & rsh_mask);
+        const auto xn_ex = (x[1] >> rsh) & rsh_mask;
+
+        const auto v = reciprocal_2by1(yn);
+        const auto res1 = udivrem_2by1({xn_hi, xn_ex}, yn, v);
+        const auto res2 = udivrem_2by1({xn_lo, res1.rem}, yn, v);
+        return {{res2.quot, res1.quot}, res2.rem >> lsh};
+    }
+
+    if (y[1] > x[1])
+        return {0, x};
+
+    const auto lsh = clz(y[1]);
+    if (lsh == 0)
+    {
+        const auto q = unsigned{y[1] < x[1]} | unsigned{y[0] <= x[0]};
+        return {q, x - (q ? y : 0)};
+    }
+
+    const auto rsh = 64 - lsh;
+
+    const auto yn_lo = y[0] << lsh;
+    const auto yn_hi = (y[1] << lsh) | (y[0] >> rsh);
+    const auto xn_lo = x[0] << lsh;
+    const auto xn_hi = (x[1] << lsh) | (x[0] >> rsh);
+    const auto xn_ex = x[1] >> rsh;
+
+    const auto v = reciprocal_3by2({yn_lo, yn_hi});
+    const auto res = udivrem_3by2(xn_ex, xn_hi, xn_lo, {yn_lo, yn_hi}, v);
+
+    return {res.quot, res.rem >> lsh};
+}
+
+inline div_result<uint128> sdivrem(uint128 x, uint128 y) noexcept
+{
+    constexpr auto sign_mask = uint128{1} << 127;
+    const auto x_is_neg = (x & sign_mask) != 0;
+    const auto y_is_neg = (y & sign_mask) != 0;
+
+    const auto x_abs = x_is_neg ? -x : x;
+    const auto y_abs = y_is_neg ? -y : y;
+
+    const auto q_is_neg = x_is_neg ^ y_is_neg;
+
+    const auto res = udivrem(x_abs, y_abs);
+
+    return {q_is_neg ? -res.quot : res.quot, x_is_neg ? -res.rem : res.rem};
+}
+
+inline uint128 operator/(uint128 x, uint128 y) noexcept
+{
+    return udivrem(x, y).quot;
+}
+
+inline uint128 operator%(uint128 x, uint128 y) noexcept
+{
+    return udivrem(x, y).rem;
+}
+
+inline uint128& operator/=(uint128& x, uint128 y) noexcept
+{
+    return x = x / y;
+}
+
+inline uint128& operator%=(uint128& x, uint128 y) noexcept
+{
+    return x = x % y;
+}
+
+/// @}
+
+}  // namespace intx
+
+
+namespace std
+{
+template <unsigned N>
+struct numeric_limits<intx::uint<N>>  // NOLINT(cert-dcl58-cpp)
+{
+    using type = intx::uint<N>;
+
+    static constexpr bool is_specialized = true;
+    static constexpr bool is_integer = true;
+    static constexpr bool is_signed = false;
+    static constexpr bool is_exact = true;
+    static constexpr bool has_infinity = false;
+    static constexpr bool has_quiet_NaN = false;
+    static constexpr bool has_signaling_NaN = false;
+    static constexpr float_denorm_style has_denorm = denorm_absent;
+    static constexpr bool has_denorm_loss = false;
+    static constexpr float_round_style round_style = round_toward_zero;
+    static constexpr bool is_iec559 = false;
+    static constexpr bool is_bounded = true;
+    static constexpr bool is_modulo = true;
+    static constexpr int digits = CHAR_BIT * sizeof(type);
+    static constexpr int digits10 = int(0.3010299956639812 * digits);
+    static constexpr int max_digits10 = 0;
+    static constexpr int radix = 2;
+    static constexpr int min_exponent = 0;
+    static constexpr int min_exponent10 = 0;
+    static constexpr int max_exponent = 0;
+    static constexpr int max_exponent10 = 0;
+    static constexpr bool traps = std::numeric_limits<unsigned>::traps;
+    static constexpr bool tinyness_before = false;
+
+    static constexpr type min() noexcept { return 0; }
+    static constexpr type lowest() noexcept { return min(); }
+    static constexpr type max() noexcept { return ~type{0}; }
+    static constexpr type epsilon() noexcept { return 0; }
+    static constexpr type round_error() noexcept { return 0; }
+    static constexpr type infinity() noexcept { return 0; }
+    static constexpr type quiet_NaN() noexcept { return 0; }
+    static constexpr type signaling_NaN() noexcept { return 0; }
+    static constexpr type denorm_min() noexcept { return 0; }
+};
+}  // namespace std
+
+namespace intx
+{
+template <typename T>
+[[noreturn]] inline void throw_(const char* what)
+{
+#if __cpp_exceptions
+    throw T{what};
+#else
+    std::fputs(what, stderr);
+    std::abort();
+#endif
+}
+
+inline constexpr int from_dec_digit(char c)
+{
+    if (c < '0' || c > '9')
+        throw_<std::invalid_argument>("invalid digit");
+    return c - '0';
+}
+
+inline constexpr int from_hex_digit(char c)
+{
+    if (c >= 'a' && c <= 'f')
+        return c - ('a' - 10);
+    if (c >= 'A' && c <= 'F')
+        return c - ('A' - 10);
+    return from_dec_digit(c);
+}
+
+template <typename Int>
+inline constexpr Int from_string(const char* str)
+{
+    auto s = str;
+    auto x = Int{};
+    int num_digits = 0;
+
+    if (s[0] == '0' && s[1] == 'x')
+    {
+        s += 2;
+        while (const auto c = *s++)
+        {
+            if (++num_digits > int{sizeof(x) * 2})
+                throw_<std::out_of_range>(str);
+            x = (x << uint64_t{4}) | from_hex_digit(c);
+        }
+        return x;
+    }
+
+    while (const auto c = *s++)
+    {
+        if (num_digits++ > std::numeric_limits<Int>::digits10)
+            throw_<std::out_of_range>(str);
+
+        const auto d = from_dec_digit(c);
+        x = x * Int{10} + d;
+        if (x < d)
+            throw_<std::out_of_range>(str);
+    }
+    return x;
+}
+
+template <typename Int>
+inline constexpr Int from_string(const std::string& s)
+{
+    return from_string<Int>(s.c_str());
+}
+
+inline constexpr uint128 operator""_u128(const char* s)
+{
+    return from_string<uint128>(s);
+}
+
+template <unsigned N>
+inline std::string to_string(uint<N> x, int base = 10)
+{
+    if (base < 2 || base > 36)
+        throw_<std::invalid_argument>("invalid base");
+
+    if (x == 0)
+        return "0";
+
+    auto s = std::string{};
+    while (x != 0)
+    {
+        // TODO: Use constexpr udivrem_1?
+        const auto res = udivrem(x, uint<N>{base});
+        const auto d = int(res.rem);
+        const auto c = d < 10 ? '0' + d : 'a' + d - 10;
+        s.push_back(char(c));
+        x = res.quot;
+    }
+    std::reverse(s.begin(), s.end());
+    return s;
+}
+
+template <unsigned N>
+inline std::string hex(uint<N> x)
+{
+    return to_string(x, 16);
+}
+
 template <unsigned N>
 struct uint
 {
@@ -66,6 +1085,97 @@ public:
         static_assert(sizeof(Int) <= sizeof(uint64_t));
         return static_cast<Int>(words_[0]);
     }
+
+    friend inline constexpr uint operator+(const uint& x, const uint& y) noexcept
+    {
+        return addc(x, y).value;
+    }
+
+    inline constexpr uint& operator+=(const uint& y) noexcept { return *this = *this + y; }
+
+    inline constexpr uint operator-() const noexcept { return ~*this + uint{1}; }
+
+    friend inline constexpr uint operator-(const uint& x, const uint& y) noexcept
+    {
+        return subc(x, y).value;
+    }
+
+    inline constexpr uint& operator-=(const uint& y) noexcept { return *this = *this - y; }
+
+    /// Multiplication implementation using word access
+    /// and discarding the high part of the result product.
+    friend inline constexpr uint operator*(const uint& x, const uint& y) noexcept
+    {
+        uint<N> p;
+        for (size_t j = 0; j < num_words; j++)
+        {
+            uint64_t k = 0;
+            for (size_t i = 0; i < (num_words - j - 1); i++)
+            {
+                auto a = addc(p[i + j], k);
+                auto t = umul(x[i], y[j]) + uint128{a.value, a.carry};
+                p[i + j] = t[0];
+                k = t[1];
+            }
+            p[num_words - 1] += x[num_words - j - 1] * y[j] + k;
+        }
+        return p;
+    }
+
+    inline constexpr uint& operator*=(const uint& y) noexcept { return *this = *this * y; }
+
+    friend inline constexpr uint operator/(const uint& x, const uint& y) noexcept
+    {
+        return udivrem(x, y).quot;
+    }
+
+    friend inline constexpr uint operator%(const uint& x, const uint& y) noexcept
+    {
+        return udivrem(x, y).rem;
+    }
+
+    inline constexpr uint& operator/=(const uint& y) noexcept { return *this = *this / y; }
+
+    inline constexpr uint& operator%=(const uint& y) noexcept { return *this = *this % y; }
+
+
+    inline constexpr uint operator~() const noexcept
+    {
+        uint z;
+        for (size_t i = 0; i < num_words; ++i)
+            z[i] = ~words_[i];
+        return z;
+    }
+
+    friend inline constexpr uint operator|(const uint& x, const uint& y) noexcept
+    {
+        uint z;
+        for (size_t i = 0; i < num_words; ++i)
+            z[i] = x[i] | y[i];
+        return z;
+    }
+
+    inline constexpr uint& operator|=(const uint& y) noexcept { return *this = *this | y; }
+
+    friend inline constexpr uint operator&(const uint& x, const uint& y) noexcept
+    {
+        uint z;
+        for (size_t i = 0; i < num_words; ++i)
+            z[i] = x[i] & y[i];
+        return z;
+    }
+
+    inline constexpr uint& operator&=(const uint& y) noexcept { return *this = *this & y; }
+
+    friend inline constexpr uint operator^(const uint& x, const uint& y) noexcept
+    {
+        uint z;
+        for (size_t i = 0; i < num_words; ++i)
+            z[i] = x[i] ^ y[i];
+        return z;
+    }
+
+    inline constexpr uint& operator^=(const uint& y) noexcept { return *this = *this ^ y; }
 };
 
 using uint192 = uint<192>;
@@ -77,10 +1187,10 @@ using uint512 = uint<512>;
 template <unsigned N>
 inline constexpr bool operator==(const uint<N>& x, const uint<N>& y) noexcept
 {
-    bool result = true;
+    uint64_t folded = 0;
     for (size_t i = 0; i < uint<N>::num_words; ++i)
-        result &= (x[i] == y[i]);
-    return result;
+        folded |= (x[i] ^ y[i]);
+    return folded == 0;
 }
 
 template <unsigned N, typename T,
@@ -121,18 +1231,21 @@ inline constexpr bool operator!=(const T& x, const uint<N>& y) noexcept
 #if !defined(_MSC_VER) || _MSC_VER < 1916  // This kills MSVC 2017 compiler.
 inline constexpr bool operator<(const uint256& x, const uint256& y) noexcept
 {
-    const auto xhi = uint128{x[2], x[3]};
-    const auto xlo = uint128{x[0], x[1]};
-    const auto yhi = uint128{y[2], y[3]};
-    const auto ylo = uint128{y[0], y[1]};
-    return (xhi < yhi) | ((xhi == yhi) & (xlo < ylo));
+    auto xp = uint128{x[2], x[3]};
+    auto yp = uint128{y[2], y[3]};
+    if (xp == yp)
+    {
+        xp = uint128{x[0], x[1]};
+        yp = uint128{y[0], y[1]};
+    }
+    return xp < yp;
 }
 #endif
 
 template <unsigned N>
 inline constexpr bool operator<(const uint<N>& x, const uint<N>& y) noexcept
 {
-    return sub_with_carry(x, y).carry;
+    return subc(x, y).carry;
 }
 
 template <unsigned N, typename T,
@@ -223,42 +1336,6 @@ inline constexpr bool slt(const uint<N>& x, const uint<N>& y) noexcept
     const auto x_neg = static_cast<int64_t>(x[top_word_idx]) < 0;
     const auto y_neg = static_cast<int64_t>(y[top_word_idx]) < 0;
     return ((x_neg ^ y_neg) != 0) ? x_neg : x < y;
-}
-
-template <unsigned N>
-inline constexpr uint<N> operator|(const uint<N>& x, const uint<N>& y) noexcept
-{
-    uint<N> z;
-    for (size_t i = 0; i < uint<N>::num_words; ++i)
-        z[i] = x[i] | y[i];
-    return z;
-}
-
-template <unsigned N>
-inline constexpr uint<N> operator&(const uint<N>& x, const uint<N>& y) noexcept
-{
-    uint<N> z;
-    for (size_t i = 0; i < uint<N>::num_words; ++i)
-        z[i] = x[i] & y[i];
-    return z;
-}
-
-template <unsigned N>
-inline constexpr uint<N> operator^(const uint<N>& x, const uint<N>& y) noexcept
-{
-    uint<N> z;
-    for (size_t i = 0; i < uint<N>::num_words; ++i)
-        z[i] = x[i] ^ y[i];
-    return z;
-}
-
-template <unsigned N>
-inline constexpr uint<N> operator~(const uint<N>& x) noexcept
-{
-    uint<N> z;
-    for (size_t i = 0; i < uint<N>::num_words; ++i)
-        z[i] = ~x[i];
-    return z;
 }
 
 
@@ -439,48 +1516,18 @@ inline constexpr const uint64_t* as_words(const uint<N>& x) noexcept
     return &x[0];
 }
 
-template <unsigned N>
-inline uint8_t* as_bytes(uint<N>& x) noexcept
+template <typename T>
+inline uint8_t* as_bytes(T& x) noexcept
 {
-    return reinterpret_cast<uint8_t*>(as_words(x));
+    static_assert(std::is_trivially_copyable_v<T>);  // As in bit_cast.
+    return reinterpret_cast<uint8_t*>(&x);
 }
 
-template <unsigned N>
-inline const uint8_t* as_bytes(const uint<N>& x) noexcept
+template <typename T>
+inline const uint8_t* as_bytes(const T& x) noexcept
 {
-    return reinterpret_cast<const uint8_t*>(as_words(x));
-}
-
-template <unsigned N>
-inline constexpr uint<N> operator+(const uint<N>& x, const uint<N>& y) noexcept
-{
-    return add_with_carry(x, y).value;
-}
-
-template <unsigned N>
-inline constexpr uint<N> operator-(const uint<N>& x) noexcept
-{
-    return ~x + uint<N>{1};
-}
-
-template <unsigned N>
-inline constexpr uint<N> operator-(const uint<N>& x, const uint<N>& y) noexcept
-{
-    return sub_with_carry(x, y).value;
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N>& operator+=(uint<N>& x, const T& y) noexcept
-{
-    return x = x + y;
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N>& operator-=(uint<N>& x, const T& y) noexcept
-{
-    return x = x - y;
+    static_assert(std::is_trivially_copyable_v<T>);  // As in bit_cast.
+    return reinterpret_cast<const uint8_t*>(&x);
 }
 
 template <unsigned N>
@@ -494,42 +1541,14 @@ inline constexpr uint<2 * N> umul(const uint<N>& x, const uint<N>& y) noexcept
         uint64_t k = 0;
         for (size_t i = 0; i < num_words; ++i)
         {
-            const auto t = umul(x[i], y[j]) + p[i + j] + k;
+            auto a = addc(p[i + j], k);
+            auto t = umul(x[i], y[j]) + uint128{a.value, a.carry};
             p[i + j] = t[0];
             k = t[1];
         }
         p[j + num_words] = k;
     }
     return p;
-}
-
-/// Multiplication implementation using word access
-/// and discarding the high part of the result product.
-template <unsigned N>
-inline constexpr uint<N> operator*(const uint<N>& x, const uint<N>& y) noexcept
-{
-    constexpr auto num_words = uint<N>::num_words;
-
-    uint<N> p;
-    for (size_t j = 0; j < num_words; j++)
-    {
-        uint64_t k = 0;
-        for (size_t i = 0; i < (num_words - j - 1); i++)
-        {
-            const auto t = umul(x[i], y[j]) + p[i + j] + k;
-            p[i + j] = t[0];
-            k = t[1];
-        }
-        p[num_words - 1] += x[num_words - j - 1] * y[j] + k;
-    }
-    return p;
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N>& operator*=(uint<N>& x, const T& y) noexcept
-{
-    return x = x * y;
 }
 
 template <unsigned N>
@@ -596,7 +1615,7 @@ inline constexpr unsigned clz_nonzero(uint64_t x) noexcept
 }
 
 template <unsigned M, unsigned N>
-struct normalized_div_args
+struct normalized_div_args  // NOLINT(cppcoreguidelines-pro-type-member-init)
 {
     uint<N> divisor;
     uint<M + 64> numerator;
@@ -609,7 +1628,6 @@ template <unsigned M, unsigned N>
 [[gnu::always_inline]] inline normalized_div_args<M, N> normalize(
     const uint<M>& numerator, const uint<N>& denominator) noexcept
 {
-    // FIXME: Make the implementation type independent
     static constexpr auto num_numerator_words = uint<M>::num_words;
     static constexpr auto num_denominator_words = uint<N>::num_words;
 
@@ -709,7 +1727,7 @@ inline bool add(uint64_t s[], const uint64_t x[], const uint64_t y[], int len) n
 
     bool carry = false;
     for (int i = 0; i < len; ++i)
-        std::tie(s[i], carry) = add_with_carry(x[i], y[i], carry);
+        std::tie(s[i], carry) = addc(x[i], y[i], carry);
     return carry;
 }
 
@@ -723,11 +1741,11 @@ inline uint64_t submul(
     uint64_t borrow = 0;
     for (int i = 0; i < len; ++i)
     {
-        const auto s = sub_with_carry(x[i], borrow);
+        const auto s = x[i] - borrow;
         const auto p = umul(y[i], multiplier);
-        const auto t = sub_with_carry(s.value, p[0]);
-        r[i] = t.value;
-        borrow = p[1] + s.carry + t.carry;
+        borrow = p[1] + (x[i] < s);
+        r[i] = s - p[0];
+        borrow += (s < r[i]);
     }
     return borrow;
 }
@@ -746,7 +1764,7 @@ inline void udivrem_knuth(
         const auto u1 = u[j + dlen - 1];
         const auto u0 = u[j + dlen - 2];
 
-        uint64_t qhat;
+        uint64_t qhat{};
         if (INTX_UNLIKELY((uint128{u1, u2}) == divisor))  // Division overflows.
         {
             qhat = ~uint64_t{0};
@@ -758,10 +1776,10 @@ inline void udivrem_knuth(
             uint128 rhat;
             std::tie(qhat, rhat) = udivrem_3by2(u2, u1, u0, divisor, reciprocal);
 
-            bool carry;
+            bool carry{};
             const auto overflow = submul(&u[j], &u[j], d, dlen - 2, qhat);
-            std::tie(u[j + dlen - 2], carry) = sub_with_carry(rhat[0], overflow);
-            std::tie(u[j + dlen - 1], carry) = sub_with_carry(rhat[1], carry);
+            std::tie(u[j + dlen - 2], carry) = subc(rhat[0], overflow);
+            std::tie(u[j + dlen - 1], carry) = subc(rhat[1], carry);
 
             if (INTX_UNLIKELY(carry))
             {
@@ -777,7 +1795,7 @@ inline void udivrem_knuth(
 }  // namespace internal
 
 template <unsigned M, unsigned N>
-div_result<uint<M>, uint<N>> udivrem(const uint<M>& u, const uint<N>& v) noexcept
+constexpr div_result<uint<M>, uint<N>> udivrem(const uint<M>& u, const uint<N>& v) noexcept
 {
     auto na = internal::normalize(u, v);
 
@@ -831,30 +1849,9 @@ inline constexpr div_result<uint<N>> sdivrem(const uint<N>& u, const uint<N>& v)
     return {q_is_neg ? -res.quot : res.quot, u_is_neg ? -res.rem : res.rem};
 }
 
-template <unsigned N>
-inline constexpr uint<N> operator/(const uint<N>& x, const uint<N>& y) noexcept
+inline constexpr uint256 bswap(const uint256& x) noexcept
 {
-    return udivrem(x, y).quot;
-}
-
-template <unsigned N>
-inline constexpr uint<N> operator%(const uint<N>& x, const uint<N>& y) noexcept
-{
-    return udivrem(x, y).rem;
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N>& operator/=(uint<N>& x, const T& y) noexcept
-{
-    return x = x / y;
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N>& operator%=(uint<N>& x, const T& y) noexcept
-{
-    return x = x % y;
+    return {bswap(x[3]), bswap(x[2]), bswap(x[1]), bswap(x[0])};
 }
 
 template <unsigned N>
@@ -869,139 +1866,6 @@ inline constexpr uint<N> bswap(const uint<N>& x) noexcept
 
 
 // Support for type conversions for binary operators.
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N> operator+(const uint<N>& x, const T& y) noexcept
-{
-    return x + uint<N>(y);
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N> operator+(const T& x, const uint<N>& y) noexcept
-{
-    return uint<N>(x) + y;
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N> operator-(const uint<N>& x, const T& y) noexcept
-{
-    return x - uint<N>(y);
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N> operator-(const T& x, const uint<N>& y) noexcept
-{
-    return uint<N>(x) - y;
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N> operator*(const uint<N>& x, const T& y) noexcept
-{
-    return x * uint<N>(y);
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N> operator*(const T& x, const uint<N>& y) noexcept
-{
-    return uint<N>(x) * y;
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N> operator/(const uint<N>& x, const T& y) noexcept
-{
-    return x / uint<N>(y);
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N> operator/(const T& x, const uint<N>& y) noexcept
-{
-    return uint<N>(x) / y;
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N> operator%(const uint<N>& x, const T& y) noexcept
-{
-    return x % uint<N>(y);
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N> operator%(const T& x, const uint<N>& y) noexcept
-{
-    return uint<N>(x) % y;
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N> operator|(const uint<N>& x, const T& y) noexcept
-{
-    return x | uint<N>(y);
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N> operator|(const T& x, const uint<N>& y) noexcept
-{
-    return uint<N>(x) | y;
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N> operator&(const uint<N>& x, const T& y) noexcept
-{
-    return x & uint<N>(y);
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N> operator&(const T& x, const uint<N>& y) noexcept
-{
-    return uint<N>(x) & y;
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N> operator^(const uint<N>& x, const T& y) noexcept
-{
-    return x ^ uint<N>(y);
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N> operator^(const T& x, const uint<N>& y) noexcept
-{
-    return uint<N>(x) ^ y;
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N>& operator|=(uint<N>& x, const T& y) noexcept
-{
-    return x = x | y;
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N>& operator&=(uint<N>& x, const T& y) noexcept
-{
-    return x = x & y;
-}
-
-template <unsigned N, typename T,
-    typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
-inline constexpr uint<N>& operator^=(uint<N>& x, const T& y) noexcept
-{
-    return x = x ^ y;
-}
 
 template <unsigned N, typename T,
     typename = typename std::enable_if<std::is_convertible<T, uint<N>>::value>::type>
@@ -1020,7 +1884,33 @@ inline constexpr uint<N>& operator>>=(uint<N>& x, const T& y) noexcept
 
 inline uint256 addmod(const uint256& x, const uint256& y, const uint256& mod) noexcept
 {
-    const auto s = add_with_carry(x, y);
+    // Fast path for mod >= 2^192, with x and y at most slightly bigger than mod.
+    // This is always the case when x and y are already reduced modulo mod.
+    // Based on https://github.com/holiman/uint256/pull/86.
+    if ((mod[3] != 0) && (x[3] <= mod[3]) && (y[3] <= mod[3]))
+    {
+        // Normalize x in case it is bigger than mod.
+        auto xn = x;
+        auto xd = subc(x, mod);
+        if (!xd.carry)
+            xn = xd.value;
+
+        // Normalize y in case it is bigger than mod.
+        auto yn = y;
+        auto yd = subc(y, mod);
+        if (!yd.carry)
+            yn = yd.value;
+
+        auto a = addc(xn, yn);
+        auto av = a.value;
+        auto b = subc(av, mod);
+        auto bv = b.value;
+        if (a.carry || !b.carry)
+            return bv;
+        return av;
+    }
+
+    auto s = addc(x, y);
     uint<256 + 64> n = s.value;
     n[4] = s.carry;
     return udivrem(n, mod).rem;
@@ -1032,49 +1922,105 @@ inline uint256 mulmod(const uint256& x, const uint256& y, const uint256& mod) no
 }
 
 
-inline constexpr uint256 operator"" _u256(const char* s) noexcept
+inline constexpr uint256 operator"" _u256(const char* s)
 {
     return from_string<uint256>(s);
 }
 
-inline constexpr uint512 operator"" _u512(const char* s) noexcept
+inline constexpr uint512 operator"" _u512(const char* s)
 {
     return from_string<uint512>(s);
 }
 
+
+/// Convert native representation to/from little-endian byte order.
+/// intx and built-in integral types are supported.
+template <typename T>
+inline constexpr T to_little_endian(const T& x) noexcept
+{
+    if constexpr (byte_order_is_little_endian)
+        return x;
+    else if constexpr (std::is_integral_v<T>)
+        return bswap(x);
+    else  // Wordwise bswap.
+    {
+        T r;
+        for (size_t i = 0; i < T::num_words; ++i)
+            r[i] = bswap(x[i]);
+        return r;
+    }
+}
+
+/// Convert native representation to/from big-endian byte order.
+/// intx and built-in integral types are supported.
+template <typename T>
+inline constexpr T to_big_endian(const T& x) noexcept
+{
+    if constexpr (byte_order_is_little_endian)
+        return bswap(x);
+    else if constexpr (std::is_integral_v<T>)
+        return x;
+    else  // Swap words.
+    {
+        T r;
+        for (size_t i = 0; i < T::num_words; ++i)
+            r[T::num_words - 1 - i] = x[i];
+        return r;
+    }
+}
+
 namespace le  // Conversions to/from LE bytes.
 {
-template <typename IntT, unsigned M>
-inline IntT load(const uint8_t (&bytes)[M]) noexcept
+template <typename T, unsigned M>
+inline T load(const uint8_t (&src)[M]) noexcept
 {
-    static_assert(M == IntT::num_bits / 8,
-        "the size of source bytes must match the size of the destination uint");
-    auto x = IntT{};
-    std::memcpy(&x, bytes, sizeof(x));
-    return x;
+    static_assert(
+        M == sizeof(T), "the size of source bytes must match the size of the destination uint");
+    T x;
+    std::memcpy(&x, src, sizeof(x));
+    return to_little_endian(x);
 }
 
-template <unsigned N>
-inline void store(uint8_t (&dst)[N / 8], const intx::uint<N>& x) noexcept
+template <typename T>
+inline void store(uint8_t (&dst)[sizeof(T)], const T& x) noexcept
 {
-    std::memcpy(dst, &x, sizeof(x));
+    const auto d = to_little_endian(x);
+    std::memcpy(dst, &d, sizeof(d));
 }
 
+namespace unsafe
+{
+template <typename T>
+inline T load(const uint8_t* src) noexcept
+{
+    T x;
+    std::memcpy(&x, src, sizeof(x));
+    return to_little_endian(x);
+}
+
+template <typename T>
+inline void store(uint8_t* dst, const T& x) noexcept
+{
+    const auto d = to_little_endian(x);
+    std::memcpy(dst, &d, sizeof(d));
+}
+}  // namespace unsafe
 }  // namespace le
 
 
 namespace be  // Conversions to/from BE bytes.
 {
-/// Loads an uint value from bytes of big-endian order.
-/// If the size of bytes is smaller than the result uint, the value is zero-extended.
-template <typename IntT, unsigned M>
-inline IntT load(const uint8_t (&bytes)[M]) noexcept
+/// Loads an integer value from bytes of big-endian order.
+/// If the size of bytes is smaller than the result, the value is zero-extended.
+template <typename T, unsigned M>
+inline T load(const uint8_t (&src)[M]) noexcept
 {
-    static_assert(M <= IntT::num_bits / 8,
+    static_assert(M <= sizeof(T),
         "the size of source bytes must not exceed the size of the destination uint");
-    auto x = IntT{};
-    std::memcpy(&as_bytes(x)[IntT::num_bits / 8 - M], bytes, M);
-    return bswap(x);
+    T x{};
+    std::memcpy(&as_bytes(x)[sizeof(T) - M], src, M);
+    x = to_big_endian(x);
+    return x;
 }
 
 template <typename IntT, typename T>
@@ -1083,20 +2029,20 @@ inline IntT load(const T& t) noexcept
     return load<IntT>(t.bytes);
 }
 
-/// Stores an uint value in a bytes array in big-endian order.
-template <unsigned N>
-inline void store(uint8_t (&dst)[N / 8], const intx::uint<N>& x) noexcept
+/// Stores an integer value in a bytes array in big-endian order.
+template <typename T>
+inline void store(uint8_t (&dst)[sizeof(T)], const T& x) noexcept
 {
-    const auto d = bswap(x);
+    const auto d = to_big_endian(x);
     std::memcpy(dst, &d, sizeof(d));
 }
 
-/// Stores an uint value in .bytes field of type T. The .bytes must be an array of uint8_t
+/// Stores an SrcT value in .bytes field of type DstT. The .bytes must be an array of uint8_t
 /// of the size matching the size of uint.
-template <typename T, unsigned N>
-inline T store(const intx::uint<N>& x) noexcept
+template <typename DstT, typename SrcT>
+inline DstT store(const SrcT& x) noexcept
 {
-    T r{};
+    DstT r{};
     store(r.bytes, x);
     return r;
 }
@@ -1105,17 +2051,16 @@ inline T store(const intx::uint<N>& x) noexcept
 /// Only the least significant bytes from big-endian representation of the uint
 /// are stored in the result bytes array up to array's size.
 template <unsigned M, unsigned N>
-inline void trunc(uint8_t (&dst)[M], const intx::uint<N>& x) noexcept
+inline void trunc(uint8_t (&dst)[M], const uint<N>& x) noexcept
 {
     static_assert(M < N / 8, "destination must be smaller than the source value");
-    const auto d = bswap(x);
-    const auto b = as_bytes(d);
-    std::memcpy(dst, &b[sizeof(d) - M], M);
+    const auto d = to_big_endian(x);
+    std::memcpy(dst, &as_bytes(d)[sizeof(d) - M], M);
 }
 
 /// Stores the truncated value of an uint in the .bytes field of an object of type T.
 template <typename T, unsigned N>
-inline T trunc(const intx::uint<N>& x) noexcept
+inline T trunc(const uint<N>& x) noexcept
 {
     T r{};
     trunc(r.bytes, x);
@@ -1127,23 +2072,50 @@ namespace unsafe
 /// Loads an uint value from a buffer. The user must make sure
 /// that the provided buffer is big enough. Therefore marked "unsafe".
 template <typename IntT>
-inline IntT load(const uint8_t* bytes) noexcept
+inline IntT load(const uint8_t* src) noexcept
 {
-    auto x = IntT{};
-    std::memcpy(&x, bytes, sizeof(x));
-    return bswap(x);
+    // Align bytes.
+    // TODO: Using memcpy() directly triggers this optimization bug in GCC:
+    //   https://gcc.gnu.org/bugzilla/show_bug.cgi?id=107837
+    alignas(IntT) std::byte aligned_storage[sizeof(IntT)];
+    std::memcpy(&aligned_storage, src, sizeof(IntT));
+    // TODO(C++23): Use std::start_lifetime_as<uint256>().
+    return to_big_endian(*reinterpret_cast<const IntT*>(&aligned_storage));
 }
 
-/// Stores an uint value at the provided pointer in big-endian order. The user must make sure
+/// Stores an integer value at the provided pointer in big-endian order. The user must make sure
 /// that the provided buffer is big enough to fit the value. Therefore marked "unsafe".
-template <unsigned N>
-inline void store(uint8_t* dst, const intx::uint<N>& x) noexcept
+template <typename T>
+inline void store(uint8_t* dst, const T& x) noexcept
 {
-    const auto d = bswap(x);
+    const auto d = to_big_endian(x);
     std::memcpy(dst, &d, sizeof(d));
 }
+
+/// Specialization for uint256.
+inline void store(uint8_t* dst, const uint256& x) noexcept
+{
+    // Store byte-swapped words in primitive temporaries. This helps with memory aliasing
+    // and GCC bug https://gcc.gnu.org/bugzilla/show_bug.cgi?id=107837
+    // TODO: Use std::byte instead of uint8_t.
+    const auto v0 = to_big_endian(x[0]);
+    const auto v1 = to_big_endian(x[1]);
+    const auto v2 = to_big_endian(x[2]);
+    const auto v3 = to_big_endian(x[3]);
+
+    // Store words in reverse (big-endian) order, write addresses are ascending.
+    std::memcpy(dst, &v3, sizeof(v3));
+    std::memcpy(dst + 8, &v2, sizeof(v2));
+    std::memcpy(dst + 16, &v1, sizeof(v1));
+    std::memcpy(dst + 24, &v0, sizeof(v0));
+}
+
 }  // namespace unsafe
 
 }  // namespace be
 
 }  // namespace intx
+
+#ifdef _MSC_VER
+    #pragma warning(pop)
+#endif
